refactor : 마커 훅 분리
